Fix formatting of certificate status in test

+++ src/tests/test_x509_path.cpp
@@ -1143,7 +1143,9 @@ class Path_Validation_With_OCSP_Tests final : public Test {
                                                                std::chrono::milliseconds(0),
                                                                {ocsp_ee});
 
-            result.test_is_eq("should result in expected validation status code", path_result.result(), expected);
+            result.test_is_eq("should result in expected validation status code",
+                              static_cast<uint32_t>(path_result.result()),
+                              static_cast<uint32_t>(expected));
             if(also_expected) {
                result.confirm("Secondary error is also present",
                               flatten(path_result.all_statuses()).contains(also_expected.value()));
